// src/types/agent.ts
export interface ToolConfig {
  id: string;
  envs: Record<string, string>;
}

export interface MCPServerConfig {
  id: string;
  envs: Record<string, string>;
  tools: string[];
  selected_tools?: string[];
}

export interface CustomMCPServer {
  url: string;
  headers?: Record<string, string>;
}
// ...
export interface CustomTools {
  http_tools: HTTPTool[];
}

export interface WorkflowData {
  nodes: any[];
  edges: any[];
}

export interface TaskConfig {
  agent_id: string;
  description: string;
  expected_output: string;
  enabled_tools?: string[];
}

export interface AgentConfig {
  // LLM config
  api_key?: string;
  tools?: ToolConfig[];
  custom_tools?: CustomTools;
  mcp_servers?: MCPServerConfig[];
  custom_mcp_servers?: CustomMCPServer[];

  // Sequential, Parallel e Loop config
  sub_agents?: string[];
  agent_tools?: string[];

  // Loop config
  max_iterations?: number;

  // Workflow config
  workflow?: WorkflowData;

  // Task config
  tasks?: TaskConfig[];
}

export interface Agent {
  id: string;
  client_id: string;
  folder_id?: string;
  name: string;
  description?: string;
  role?: string;
  goal?: string;
  type: AgentType;
  model?: string;
  api_key_id?: string;
  instruction?: string;
  agent_card_url?: string;
  config?: AgentConfig;
  tasks?: TaskConfig[];
  created_at: string;
  updated_at?: string;
}